Update __init__.py
Error with Scale value - invalid literal for int() with base 10

+++ cbpi4-hx711-loadcell/__init__.py
@@ -34,7 +34,7 @@ class CustomSensor(CBPiSensor):
         self.gain = int(self.props.get("gain",128))
         self.Interval = int(self.props.get("Interval",2))
         self.offset = int(self.props.get("offset",0))
-        self.scale = int(self.props.get("scale",1))
+        self.scale = float(self.props.get("scale",1))
         self.calibration_active = False
         self.measurement_is_running = False
